requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
chromadb
python-dotenv
streamlit
